Fixing numerical problem of compute_mllr and mllr_solve.  There are small numerical differences between the inputs typed with float64 or float32. In terms of (<6 signficiant digits).  This small difference will translate to perceivable numerical difference in the final matrix. (>5 significant digits).  This fix also marks a stable release for mllr_solve.  I also disallow user to use -regmat in norm because it is highly dangerous and known to be slow and didn't help too much.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/SphinxTrain@3731 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/s3/mllr.h
@@ -53,8 +53,8 @@ int32 regmat_read (
 
 
 int32 compute_mllr (
-             float32 *****regl,
-             float32 ****regr,
+             float64 *****regl,
+             float64 ****regr,
              const uint32  *veclen,
              uint32  nclass,
              uint32  nfeat,
@@ -71,8 +71,8 @@ invert(float32 **ainv,
        int32 len);
 
 int32 solve  (
-        float32 **regl,
-        float32 *regr,
+        float64 **regl,
+        float64 *regr,
         int32   len,
         float64 *A
        );
